pyproject.toml
[tool.poetry]
name = "django-query-view"
version = "0.1.0"
description = "Django view for querying data"
authors = ["Enrico Barzetti <enricobarzetti@gmail.com>"]
readme = "README.rst"
homepage = "https://github.com/enricobarzetti/django-query-view"
license = "mit"

[tool.poetry.dependencies]
python = "^3.8"
django-taggit = "^1.3.0"
taggit-selectize = "^2.10.0"
django-filter = "^2.4.0"
django-extensions = "^3.1.0"
furl = "^2.1.0"
django-crispy-forms = "^1.10.0"
django_jinja = "^2.7.0"
django_jinja_helpers = "^0.1.4"

[tool.poetry.dev-dependencies]
ipython = "^7.19.0"
isort = "^5.6.4"
pudb = "^2020.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
